Moving label generation to own function
