+++ src/xDate.js
@@ -6,14 +6,12 @@
 import * as d3 from 'd3';
 import {
     convertToPointScale,
+    generateDateTickValues,
+    generateLabels,
     getAxis,
     getBandWidth,
-    getDefaultXAxisLabel,
     getTimeTickFormat,
-    getTimeTicks,
     getTimeTicksMinor,
-    getXVertical,
-    getXHorizontal,
     setLabelIds,
 } from './utils';
 
@@ -46,62 +44,31 @@ export default function xaxisDate() {
         const plotWidth = plotDim[0];
         const plotHeight = plotDim[1];
         const xAxis = getAxis(align);
+
         if (intraday) {
             scale = convertToPointScale(scale);
-            xAxis
-                .tickSize(tickSize)
-                .tickFormat(getTimeTickFormat(interval, { fullYear, scale }))
-                .scale(scale);
-            xAxis.tickValues(
-                scale.domain().filter((d, i) => {
-                    let checkDate;
-                    if (i === 0) {
-                        return d.getDay();
-                    }
-                    if (i > 0) {
-                        checkDate = new Date(scale.domain()[i - 1]);
-                    }
-                    return d.getDay() !== checkDate.getDay();
-                }),
-            );
-        } else {
-            xAxis
-                .tickSize(tickSize)
-                // .ticks(getTimeTicks(interval))
-                .tickFormat(getTimeTickFormat(interval, { fullYear, scale }))
-                .scale(scale);
-            let newTicks = scale.ticks(getTimeTicks(interval));
-            const dayCheck = scale.domain()[0].getDate();
-            const monthCheck = scale.domain()[0].getMonth();
-            if (dayCheck !== 1 && monthCheck !== 0) {
-                newTicks.unshift(scale.domain()[0]);
-            }
-            if (
-                interval === 'lustrum' ||
-                interval === 'decade' ||
-                interval === 'jubilee' ||
-                interval === 'century'
-            ) {
-                newTicks.push(d3.timeYear(scale.domain()[1]));
-            }
-            if (endTicks) {
-                newTicks = scale.domain();
-            }
-            xAxis.tickValues(newTicks);
         }
 
-        const xMinor = getAxis(align);
-        if (intraday) {
-            xMinor
-                .tickSize(minorTickSize)
-                .tickFormat('')
-                .scale(scale);
-        } else {
-            xMinor
-                .tickSize(minorTickSize)
-                .ticks(getTimeTicksMinor(interval))
-                .tickFormat('')
-                .scale(scale);
+        xAxis
+            .tickSize(tickSize)
+            .tickFormat(getTimeTickFormat(interval, { fullYear, scale }))
+            .scale(scale)
+            .tickValues(
+                generateDateTickValues({
+                    intraday,
+                    scale,
+                    interval,
+                    endTicks,
+                }),
+            );
+
+        const xMinor = getAxis(align)
+            .tickSize(minorTickSize)
+            .tickFormat('')
+            .scale(scale);
+
+        if (!intraday) {
+            xMinor.ticks(getTimeTicksMinor(interval));
         }
 
         if (tickValues) {
@@ -141,37 +108,15 @@ export default function xaxisDate() {
         }
 
         if (label) {
-            const defaultLabel = getDefaultXAxisLabel(label);
-
-            const axisLabel = parent.append('g').attr('class', 'axis xAxis');
-
-            axisLabel
-                .append('text')
-                .attr(
-                    'y',
-                    getXVertical({
-                        align,
-                        vert: defaultLabel.vert,
-                        plotHeight,
-                        rem,
-                        tickSize,
-                    }),
-                )
-                .attr(
-                    'x',
-                    getXHorizontal({ hori: defaultLabel.hori, plotWidth }),
-                )
-                .text(defaultLabel.tag);
-
-            const text = axisLabel.selectAll('text');
-            const width = text.node().getBBox().width / 2;
-            const height = text.node().getBBox().height / 2;
-            const textX = text.node().getBBox().x + width;
-            const textY = text.node().getBBox().y + height;
-            text.attr(
-                'transform',
-                `rotate(${defaultLabel.rotate}, ${textX}, ${textY})`,
-            ).style('text-anchor', defaultLabel.anchor);
+            generateLabels('x', {
+                align,
+                label,
+                parent,
+                plotHeight,
+                plotWidth,
+                rem,
+                tickSize,
+            });
         }
         if (banding) {
             let bands = xAxis.tickValues();
